refactor(resources): use job to find primes concurrently

+++ src/main/java/io/inhibitor/behaviors/PrimeBehavior.java
@@ -2,16 +2,16 @@ package io.inhibitor.behaviors;
 
 public class PrimeBehavior {
 
-  public int findSmallestPrimeAfter(int bound) {
-    for (int i = bound + 1; ; i++) {
+  public long findSmallestPrimeAfter(long bound) {
+    for (long i = bound + 1; ; i++) {
       if (isPrime(i)) {
         return i;
       }
     }
   }
 
-  private boolean isPrime(int number) {
-    for (int i = 2; i * i <= number; i++) {
+  private boolean isPrime(long number) {
+    for (long i = 2; i * i <= number; i++) {
       if (number % i == 0) {
         return false;
       }
